Add tests for the overarching parse_recipe function Some of these tests test real recipes from real pdfs Added new test cases to read_text, list_instructions, list_ingredients, and get_headers, to account for cases found in real recipes Performed small bug fix in get_headers.py Made significant changes to extract_text_info in pdf_scraper.py, to make code more readable and to account for additional cases related to span spacing Updated list_instructions function to ensure numbered prefixes are only removed if in order Other bug fixes and updates to parse_recipe function to simplify and make more efficient Moved all tests into a 'test' folder, allowing them to be run all together
